feat(schema): add gmail types + env
