+++ packages/schema/src/backend/env.ts
@@ -11,6 +11,8 @@ export const BackendEnvSchema = object({
   AXIOM_DATASET: string(),
   AXIOM_ORG_ID: string(),
   AXIOM_URL: string(),
+  GOOGLE_PROVIDER_CLIENT_SECRET: string(),
+  GOOGLE_PROVIDER_CLIENT_ID: string(),
 });
 
 export type BackendEnvType = Output<typeof BackendEnvSchema>;
