+++ packages/schema/src/emailProvider/oauth.ts
@@ -1,5 +1,5 @@
 import type { Output } from "valibot";
-import { boolean, enumType, number, object, string } from "valibot";
+import { enumType, number, object, string } from "valibot";
 
 export const oauth2TokenResponseSchema = object({
   access_token: string(),
@@ -19,20 +19,3 @@ export const oauthOnboardingSchema = object({
 });
 
 export type oauthOnboardingType = Output<typeof oauthOnboardingSchema>;
-
-export const gmailProviderIDTokenSchema = object({
-  iss: string(),
-  azp: string(),
-  aud: string(),
-  sub: string(),
-  email: string(),
-  email_verified: boolean(),
-  at_hash: string(),
-  name: string(),
-  picture: string(),
-  given_name: string(),
-  family_name: string(),
-  locale: string(),
-  iat: number(),
-  exp: number(),
-});
